refactor(tasks): Migrate from object-oriented to functional task management API.
